fixup! refactor(inter-protocol): factor out startEconCommittee.js
